add error checking

+++ dcc/certificates.sh
@@ -1,21 +1,31 @@
 #!/bin/bash -v
 
-set -e
-
 # temporary script to copy certificates into place until we can use swarm
 
-cp /run/secrets/shibboleth_sp_encrypt_cert /etc/shibboleth/sp-signing-cert.pem
-cp /run/secrets/shibboleth_sp_encrypt_privkey /etc/shibboleth/sp-signing-key.pem
+set -e
+
+if [ ! -f  /etc/shibboleth/sp-signing-cert.pem ] ; then
+  cp /run/secrets/shibboleth_sp_encrypt_cert /etc/shibboleth/sp-signing-cert.pem
+fi
+if [ ! -f /etc/shibboleth/sp-signing-key.pem ] ; then
+  cp /run/secrets/shibboleth_sp_encrypt_privkey /etc/shibboleth/sp-signing-key.pem
+fi
 chown shibd:shibd /etc/shibboleth/sp-signing-cert.pem /etc/shibboleth/sp-signing-key.pem
 chmod 0400 /etc/shibboleth/sp-signing-key.pem
 chmod 0444 /etc/shibboleth/sp-signing-cert.pem
-ln -s /etc/shibboleth/sp-signing-cert.pem /etc/shibboleth/sp-encrypt-cert.pem
-ln -s /etc/shibboleth/sp-signing-key.pem /etc/shibboleth/sp-encrypt-key.pem
+ln -sf /etc/shibboleth/sp-signing-cert.pem /etc/shibboleth/sp-encrypt-cert.pem
+ln -sf /etc/shibboleth/sp-signing-key.pem /etc/shibboleth/sp-encrypt-key.pem
 
-cp /run/secrets/https_cert_file /etc/httpd/x509-certs/${DCC_INSTANCE}.pem
-cp /run/secrets/https_privkey_file /etc/httpd/x509-certs/${DCC_INSTANCE}.key
-cp /run/secrets/https_chain_file /etc/httpd/x509-certs/${DCC_INSTANCE}.cert
-chown 0444 /etc/httpd/x509-certs/${DCC_INSTANCE}.pem /etc/httpd/x509-certs/${DCC_INSTANCE}.cert
-chown 0400 /etc/httpd/x509-certs/${DCC_INSTANCE}.key
+if [ ! -f /etc/httpd/x509-certs/${DCC_INSTANCE}.pem ] ; then
+  cp /run/secrets/https_cert_file /etc/httpd/x509-certs/${DCC_INSTANCE}.pem
+fi
+if [ ! -f /etc/httpd/x509-certs/${DCC_INSTANCE}.key ] ; then
+  cp /run/secrets/https_privkey_file /etc/httpd/x509-certs/${DCC_INSTANCE}.key
+fi
+if [ ! -f /etc/httpd/x509-certs/${DCC_INSTANCE}.cert ] ; then
+  cp /run/secrets/https_chain_file /etc/httpd/x509-certs/${DCC_INSTANCE}.cert
+fi
+chmod 0444 /etc/httpd/x509-certs/${DCC_INSTANCE}.pem /etc/httpd/x509-certs/${DCC_INSTANCE}.cert
+chmod 0400 /etc/httpd/x509-certs/${DCC_INSTANCE}.key
 
 exit 0
